Expand run preprocessing to encode smiles strings as selfies

+++ scripts/run_preprocessing.py
@@ -1,12 +1,13 @@
 import argparse
 from pathlib import Path
 from shutil import rmtree
-from time import sleep
+from typing import Union
 
 import dask.dataframe as dd
 from dask.distributed import Client
 
-from mol_gen.preprocessing.preprocessor import process_dataframe
+from mol_gen.preprocessing.convert import encode_smiles_as_selfies
+from mol_gen.preprocessing.preprocessor import preprocess_smiles_dataframe
 
 
 def parse_args() -> argparse.Namespace:
@@ -31,36 +32,64 @@ def parse_args() -> argparse.Namespace:
     return args
 
 
-def main():
-    args = parse_args()
-
+def convert_and_filter_smiles_strings(
+    input_path: Union[str, Path],
+    output_path: Union[str, Path],
+    config_path: Union[str, Path],
+) -> None:
     print("Setting up dask client")
-
     with Client() as client:
         print(client.dashboard_link)
-
-        df = dd.read_parquet(args.input)
-
-        # Persist preprocessed results in intermediate directory
-        intermediate_dir = Path(args.output).joinpath("intermediate")
-
         print("Executing preprocessing steps on input molecules")
-        df.repartition(partition_size="25MB").map_partitions(
-            process_dataframe, args.config, meta=df
-        ).to_parquet(intermediate_dir)
 
-    # Drop duplicates and repartition intermediate results
-    # Persist results in output directory and wipe intermediate directory
+        df = dd.read_parquet(input_path)
+        df.repartition(partition_size="25MB").map_partitions(
+            preprocess_smiles_dataframe, config_path, meta=df
+        ).to_parquet(output_path)
+
+
+def drop_duplicates_and_repartition(input_path: str, output_path: str) -> None:
     print("Removing duplicate molecules and repartitioning files")
-    df = dd.read_parquet(intermediate_dir)
+    df = dd.read_parquet(input_path)
 
     df.drop_duplicates(subset="SMILES", split_out=df.npartitions).repartition(
         partition_size="100MB"
-    ).to_parquet(args.output)
+    ).to_parquet(output_path)
 
-    print("Removing intermediate files")
-    sleep(10)
-    rmtree(intermediate_dir)
+
+def create_selfies_from_smiles(input_path: str, output_path: str):
+    print("Setting up dask client")
+    with Client() as client:
+        print(client.dashboard_link)
+        print("Creating SELFIES from SMILES strings")
+
+        df = dd.read_parquet(input_path)
+
+        df["SELFIES"] = df["SMILES"].apply(encode_smiles_as_selfies, meta=(None, str))
+        df[["SELFIES"]].dropna().to_parquet(output_path)
+
+
+def main():
+    args = parse_args()
+
+    output_dir = Path(args.output)
+    intermediate_dir = output_dir / "intermediate"
+    preprocessed_smiles_dir = output_dir / "smiles"
+    selfies_dir = output_dir / "selfies"
+
+    if not preprocessed_smiles_dir.exists():
+        print("Starting preprocessing of SMILES strings")
+        convert_and_filter_smiles_strings(args.input, intermediate_dir, args.config)
+        drop_duplicates_and_repartition(intermediate_dir, preprocessed_smiles_dir)
+        rmtree(intermediate_dir)
+
+    else:
+        print("Skipping preprocessing of SMILES strings")
+        print("SMILES directory already exists in output directory")
+
+    if not selfies_dir.exists():
+        print("Starting encoding of SMILES strings as SELFIES")
+        create_selfies_from_smiles(preprocessed_smiles_dir, selfies_dir)
 
 
 if __name__ == "__main__":
